+++ sdk/instruction/src/error.rs
@@ -1,0 +1,338 @@
+#[cfg(feature = "frozen-abi")]
+use solana_frozen_abi_macro::{AbiEnumVisitor, AbiExample};
+use {
+    num_traits::ToPrimitive,
+    serde_derive::{Deserialize, Serialize},
+    thiserror::Error,
+};
+
+/// Builtin return values occupy the upper 32 bits
+const BUILTIN_BIT_SHIFT: usize = 32;
+macro_rules! to_builtin {
+    ($error:expr) => {
+        ($error as u64) << BUILTIN_BIT_SHIFT
+    };
+}
+
+pub const CUSTOM_ZERO: u64 = to_builtin!(1);
+pub const INVALID_ARGUMENT: u64 = to_builtin!(2);
+pub const INVALID_INSTRUCTION_DATA: u64 = to_builtin!(3);
+pub const INVALID_ACCOUNT_DATA: u64 = to_builtin!(4);
+pub const ACCOUNT_DATA_TOO_SMALL: u64 = to_builtin!(5);
+pub const INSUFFICIENT_FUNDS: u64 = to_builtin!(6);
+pub const INCORRECT_PROGRAM_ID: u64 = to_builtin!(7);
+pub const MISSING_REQUIRED_SIGNATURES: u64 = to_builtin!(8);
+pub const ACCOUNT_ALREADY_INITIALIZED: u64 = to_builtin!(9);
+pub const UNINITIALIZED_ACCOUNT: u64 = to_builtin!(10);
+pub const NOT_ENOUGH_ACCOUNT_KEYS: u64 = to_builtin!(11);
+pub const ACCOUNT_BORROW_FAILED: u64 = to_builtin!(12);
+pub const MAX_SEED_LENGTH_EXCEEDED: u64 = to_builtin!(13);
+pub const INVALID_SEEDS: u64 = to_builtin!(14);
+pub const BORSH_IO_ERROR: u64 = to_builtin!(15);
+pub const ACCOUNT_NOT_RENT_EXEMPT: u64 = to_builtin!(16);
+pub const UNSUPPORTED_SYSVAR: u64 = to_builtin!(17);
+pub const ILLEGAL_OWNER: u64 = to_builtin!(18);
+pub const MAX_ACCOUNTS_DATA_ALLOCATIONS_EXCEEDED: u64 = to_builtin!(19);
+pub const INVALID_ACCOUNT_DATA_REALLOC: u64 = to_builtin!(20);
+pub const MAX_INSTRUCTION_TRACE_LENGTH_EXCEEDED: u64 = to_builtin!(21);
+pub const BUILTIN_PROGRAMS_MUST_CONSUME_COMPUTE_UNITS: u64 = to_builtin!(22);
+pub const INVALID_ACCOUNT_OWNER: u64 = to_builtin!(23);
+pub const ARITHMETIC_OVERFLOW: u64 = to_builtin!(24);
+pub const IMMUTABLE: u64 = to_builtin!(25);
+pub const INCORRECT_AUTHORITY: u64 = to_builtin!(26);
+// Warning: Any new error codes added here must also be:
+// - Added to the below conversions
+// - Added as an equivalent to ProgramError and InstructionError
+// - Be featureized in the BPF loader to return `InstructionError::InvalidError`
+//   until the feature is activated
+
+/// Reasons the runtime might have rejected an instruction.
+///
+/// Members of this enum must not be removed, but new ones can be added.
+/// Also, it is crucial that meta-information if any that comes along with
+/// an error be consistent across software versions.  For example, it is
+/// dangerous to include error strings from 3rd party crates because they could
+/// change at any time and changes to them are difficult to detect.
+#[cfg_attr(feature = "frozen-abi", derive(AbiExample, AbiEnumVisitor))]
+#[derive(Serialize, Deserialize, Debug, Error, PartialEq, Eq, Clone)]
+pub enum InstructionError {
+    /// Deprecated! Use CustomError instead!
+    /// The program instruction returned an error
+    #[error("generic instruction error")]
+    GenericError,
+
+    /// The arguments provided to a program were invalid
+    #[error("invalid program argument")]
+    InvalidArgument,
+
+    /// An instruction's data contents were invalid
+    #[error("invalid instruction data")]
+    InvalidInstructionData,
+
+    /// An account's data contents was invalid
+    #[error("invalid account data for instruction")]
+    InvalidAccountData,
+
+    /// An account's data was too small
+    #[error("account data too small for instruction")]
+    AccountDataTooSmall,
+
+    /// An account's balance was too small to complete the instruction
+    #[error("insufficient funds for instruction")]
+    InsufficientFunds,
+
+    /// The account did not have the expected program id
+    #[error("incorrect program id for instruction")]
+    IncorrectProgramId,
+
+    /// A signature was required but not found
+    #[error("missing required signature for instruction")]
+    MissingRequiredSignature,
+
+    /// An initialize instruction was sent to an account that has already been initialized.
+    #[error("instruction requires an uninitialized account")]
+    AccountAlreadyInitialized,
+
+    /// An attempt to operate on an account that hasn't been initialized.
+    #[error("instruction requires an initialized account")]
+    UninitializedAccount,
+
+    /// Program's instruction lamport balance does not equal the balance after the instruction
+    #[error("sum of account balances before and after instruction do not match")]
+    UnbalancedInstruction,
+
+    /// Program illegally modified an account's program id
+    #[error("instruction illegally modified the program id of an account")]
+    ModifiedProgramId,
+
+    /// Program spent the lamports of an account that doesn't belong to it
+    #[error("instruction spent from the balance of an account it does not own")]
+    ExternalAccountLamportSpend,
+
+    /// Program modified the data of an account that doesn't belong to it
+    #[error("instruction modified data of an account it does not own")]
+    ExternalAccountDataModified,
+
+    /// Read-only account's lamports modified
+    #[error("instruction changed the balance of a read-only account")]
+    ReadonlyLamportChange,
+
+    /// Read-only account's data was modified
+    #[error("instruction modified data of a read-only account")]
+    ReadonlyDataModified,
+
+    /// An account was referenced more than once in a single instruction
+    // Deprecated, instructions can now contain duplicate accounts
+    #[error("instruction contains duplicate accounts")]
+    DuplicateAccountIndex,
+
+    /// Executable bit on account changed, but shouldn't have
+    #[error("instruction changed executable bit of an account")]
+    ExecutableModified,
+
+    /// Rent_epoch account changed, but shouldn't have
+    #[error("instruction modified rent epoch of an account")]
+    RentEpochModified,
+
+    /// The instruction expected additional account keys
+    #[error("insufficient account keys for instruction")]
+    NotEnoughAccountKeys,
+
+    /// Program other than the account's owner changed the size of the account data
+    #[error("program other than the account's owner changed the size of the account data")]
+    AccountDataSizeChanged,
+
+    /// The instruction expected an executable account
+    #[error("instruction expected an executable account")]
+    AccountNotExecutable,
+
+    /// Failed to borrow a reference to account data, already borrowed
+    #[error("instruction tries to borrow reference for an account which is already borrowed")]
+    AccountBorrowFailed,
+
+    /// Account data has an outstanding reference after a program's execution
+    #[error("instruction left account with an outstanding borrowed reference")]
+    AccountBorrowOutstanding,
+
+    /// The same account was multiply passed to an on-chain program's entrypoint, but the program
+    /// modified them differently.  A program can only modify one instance of the account because
+    /// the runtime cannot determine which changes to pick or how to merge them if both are modified
+    #[error("instruction modifications of multiply-passed account differ")]
+    DuplicateAccountOutOfSync,
+
+    /// Allows on-chain programs to implement program-specific error types and see them returned
+    /// by the Solana runtime. A program-specific error may be any type that is represented as
+    /// or serialized to a u32 integer.
+    #[error("custom program error: {0:#x}")]
+    Custom(u32),
+
+    /// The return value from the program was invalid.  Valid errors are either a defined builtin
+    /// error value or a user-defined error in the lower 32 bits.
+    #[error("program returned invalid error code")]
+    InvalidError,
+
+    /// Executable account's data was modified
+    #[error("instruction changed executable accounts data")]
+    ExecutableDataModified,
+
+    /// Executable account's lamports modified
+    #[error("instruction changed the balance of an executable account")]
+    ExecutableLamportChange,
+
+    /// Executable accounts must be rent exempt
+    #[error("executable accounts must be rent exempt")]
+    ExecutableAccountNotRentExempt,
+
+    /// Unsupported program id
+    #[error("Unsupported program id")]
+    UnsupportedProgramId,
+
+    /// Cross-program invocation call depth too deep
+    #[error("Cross-program invocation call depth too deep")]
+    CallDepth,
+
+    /// An account required by the instruction is missing
+    #[error("An account required by the instruction is missing")]
+    MissingAccount,
+
+    /// Cross-program invocation reentrancy not allowed for this instruction
+    #[error("Cross-program invocation reentrancy not allowed for this instruction")]
+    ReentrancyNotAllowed,
+
+    /// Length of the seed is too long for address generation
+    #[error("Length of the seed is too long for address generation")]
+    MaxSeedLengthExceeded,
+
+    /// Provided seeds do not result in a valid address
+    #[error("Provided seeds do not result in a valid address")]
+    InvalidSeeds,
+
+    /// Failed to reallocate account data of this length
+    #[error("Failed to reallocate account data")]
+    InvalidRealloc,
+
+    /// Computational budget exceeded
+    #[error("Computational budget exceeded")]
+    ComputationalBudgetExceeded,
+
+    /// Cross-program invocation with unauthorized signer or writable account
+    #[error("Cross-program invocation with unauthorized signer or writable account")]
+    PrivilegeEscalation,
+
+    /// Failed to create program execution environment
+    #[error("Failed to create program execution environment")]
+    ProgramEnvironmentSetupFailure,
+
+    /// Program failed to complete
+    #[error("Program failed to complete")]
+    ProgramFailedToComplete,
+
+    /// Program failed to compile
+    #[error("Program failed to compile")]
+    ProgramFailedToCompile,
+
+    /// Account is immutable
+    #[error("Account is immutable")]
+    Immutable,
+
+    /// Incorrect authority provided
+    #[error("Incorrect authority provided")]
+    IncorrectAuthority,
+
+    /// Failed to serialize or deserialize account data
+    ///
+    /// Warning: This error should never be emitted by the runtime.
+    ///
+    /// This error includes strings from the underlying 3rd party Borsh crate
+    /// which can be dangerous because the error strings could change across
+    /// Borsh versions. Only programs can use this error because they are
+    /// consistent across Solana software versions.
+    ///
+    #[error("Failed to serialize or deserialize account data: {0}")]
+    BorshIoError(String),
+
+    /// An account does not have enough lamports to be rent-exempt
+    #[error("An account does not have enough lamports to be rent-exempt")]
+    AccountNotRentExempt,
+
+    /// Invalid account owner
+    #[error("Invalid account owner")]
+    InvalidAccountOwner,
+
+    /// Program arithmetic overflowed
+    #[error("Program arithmetic overflowed")]
+    ArithmeticOverflow,
+
+    /// Unsupported sysvar
+    #[error("Unsupported sysvar")]
+    UnsupportedSysvar,
+
+    /// Illegal account owner
+    #[error("Provided owner is not allowed")]
+    IllegalOwner,
+
+    /// Accounts data allocations exceeded the maximum allowed per transaction
+    #[error("Accounts data allocations exceeded the maximum allowed per transaction")]
+    MaxAccountsDataAllocationsExceeded,
+
+    /// Max accounts exceeded
+    #[error("Max accounts exceeded")]
+    MaxAccountsExceeded,
+
+    /// Max instruction trace length exceeded
+    #[error("Max instruction trace length exceeded")]
+    MaxInstructionTraceLengthExceeded,
+
+    /// Builtin programs must consume compute units
+    #[error("Builtin programs must consume compute units")]
+    BuiltinProgramsMustConsumeComputeUnits,
+    // Note: For any new error added here an equivalent ProgramError and its
+    // conversions must also be added
+}
+
+impl<T> From<T> for InstructionError
+where
+    T: ToPrimitive,
+{
+    fn from(error: T) -> Self {
+        let error = error.to_u64().unwrap_or(0xbad_c0de);
+        match error {
+            CUSTOM_ZERO => Self::Custom(0),
+            INVALID_ARGUMENT => Self::InvalidArgument,
+            INVALID_INSTRUCTION_DATA => Self::InvalidInstructionData,
+            INVALID_ACCOUNT_DATA => Self::InvalidAccountData,
+            ACCOUNT_DATA_TOO_SMALL => Self::AccountDataTooSmall,
+            INSUFFICIENT_FUNDS => Self::InsufficientFunds,
+            INCORRECT_PROGRAM_ID => Self::IncorrectProgramId,
+            MISSING_REQUIRED_SIGNATURES => Self::MissingRequiredSignature,
+            ACCOUNT_ALREADY_INITIALIZED => Self::AccountAlreadyInitialized,
+            UNINITIALIZED_ACCOUNT => Self::UninitializedAccount,
+            NOT_ENOUGH_ACCOUNT_KEYS => Self::NotEnoughAccountKeys,
+            ACCOUNT_BORROW_FAILED => Self::AccountBorrowFailed,
+            MAX_SEED_LENGTH_EXCEEDED => Self::MaxSeedLengthExceeded,
+            INVALID_SEEDS => Self::InvalidSeeds,
+            BORSH_IO_ERROR => Self::BorshIoError("Unknown".to_string()),
+            ACCOUNT_NOT_RENT_EXEMPT => Self::AccountNotRentExempt,
+            UNSUPPORTED_SYSVAR => Self::UnsupportedSysvar,
+            ILLEGAL_OWNER => Self::IllegalOwner,
+            MAX_ACCOUNTS_DATA_ALLOCATIONS_EXCEEDED => Self::MaxAccountsDataAllocationsExceeded,
+            INVALID_ACCOUNT_DATA_REALLOC => Self::InvalidRealloc,
+            MAX_INSTRUCTION_TRACE_LENGTH_EXCEEDED => Self::MaxInstructionTraceLengthExceeded,
+            BUILTIN_PROGRAMS_MUST_CONSUME_COMPUTE_UNITS => {
+                Self::BuiltinProgramsMustConsumeComputeUnits
+            }
+            INVALID_ACCOUNT_OWNER => Self::InvalidAccountOwner,
+            ARITHMETIC_OVERFLOW => Self::ArithmeticOverflow,
+            IMMUTABLE => Self::Immutable,
+            INCORRECT_AUTHORITY => Self::IncorrectAuthority,
+            _ => {
+                // A valid custom error has no bits set in the upper 32
+                if error >> BUILTIN_BIT_SHIFT == 0 {
+                    Self::Custom(error as u32)
+                } else {
+                    Self::InvalidError
+                }
+            }
+        }
+    }
+}
