+++ sdk/instruction/src/stubs.rs
@@ -1,0 +1,8 @@
+use crate::Instruction;
+
+pub fn sol_get_processed_sibling_instruction(_index: usize) -> Option<Instruction> {
+    None
+}
+pub fn sol_get_stack_height() -> u64 {
+    0
+}
