+++ sdk/program/src/program_error.rs
@@ -3,9 +3,19 @@
 #![allow(clippy::arithmetic_side_effects)]
 #[cfg(feature = "borsh")]
 use borsh::io::Error as BorshIoError;
+pub use solana_instruction::error::{
+    ACCOUNT_ALREADY_INITIALIZED, ACCOUNT_BORROW_FAILED, ACCOUNT_DATA_TOO_SMALL,
+    ACCOUNT_NOT_RENT_EXEMPT, ARITHMETIC_OVERFLOW, BORSH_IO_ERROR,
+    BUILTIN_PROGRAMS_MUST_CONSUME_COMPUTE_UNITS, CUSTOM_ZERO, ILLEGAL_OWNER, IMMUTABLE,
+    INCORRECT_AUTHORITY, INCORRECT_PROGRAM_ID, INSUFFICIENT_FUNDS, INVALID_ACCOUNT_DATA,
+    INVALID_ACCOUNT_DATA_REALLOC, INVALID_ACCOUNT_OWNER, INVALID_ARGUMENT,
+    INVALID_INSTRUCTION_DATA, INVALID_SEEDS, MAX_ACCOUNTS_DATA_ALLOCATIONS_EXCEEDED,
+    MAX_INSTRUCTION_TRACE_LENGTH_EXCEEDED, MAX_SEED_LENGTH_EXCEEDED, MISSING_REQUIRED_SIGNATURES,
+    NOT_ENOUGH_ACCOUNT_KEYS, UNINITIALIZED_ACCOUNT, UNSUPPORTED_SYSVAR,
+};
 use {
     crate::{instruction::InstructionError, msg, pubkey::PubkeyError},
-    num_traits::{FromPrimitive, ToPrimitive},
+    num_traits::FromPrimitive,
     solana_decode_error::DecodeError,
     std::convert::TryFrom,
     thiserror::Error,
@@ -124,46 +134,6 @@ impl PrintProgramError for ProgramError {
         }
     }
 }
-
-/// Builtin return values occupy the upper 32 bits
-const BUILTIN_BIT_SHIFT: usize = 32;
-macro_rules! to_builtin {
-    ($error:expr) => {
-        ($error as u64) << BUILTIN_BIT_SHIFT
-    };
-}
-
-pub const CUSTOM_ZERO: u64 = to_builtin!(1);
-pub const INVALID_ARGUMENT: u64 = to_builtin!(2);
-pub const INVALID_INSTRUCTION_DATA: u64 = to_builtin!(3);
-pub const INVALID_ACCOUNT_DATA: u64 = to_builtin!(4);
-pub const ACCOUNT_DATA_TOO_SMALL: u64 = to_builtin!(5);
-pub const INSUFFICIENT_FUNDS: u64 = to_builtin!(6);
-pub const INCORRECT_PROGRAM_ID: u64 = to_builtin!(7);
-pub const MISSING_REQUIRED_SIGNATURES: u64 = to_builtin!(8);
-pub const ACCOUNT_ALREADY_INITIALIZED: u64 = to_builtin!(9);
-pub const UNINITIALIZED_ACCOUNT: u64 = to_builtin!(10);
-pub const NOT_ENOUGH_ACCOUNT_KEYS: u64 = to_builtin!(11);
-pub const ACCOUNT_BORROW_FAILED: u64 = to_builtin!(12);
-pub const MAX_SEED_LENGTH_EXCEEDED: u64 = to_builtin!(13);
-pub const INVALID_SEEDS: u64 = to_builtin!(14);
-pub const BORSH_IO_ERROR: u64 = to_builtin!(15);
-pub const ACCOUNT_NOT_RENT_EXEMPT: u64 = to_builtin!(16);
-pub const UNSUPPORTED_SYSVAR: u64 = to_builtin!(17);
-pub const ILLEGAL_OWNER: u64 = to_builtin!(18);
-pub const MAX_ACCOUNTS_DATA_ALLOCATIONS_EXCEEDED: u64 = to_builtin!(19);
-pub const INVALID_ACCOUNT_DATA_REALLOC: u64 = to_builtin!(20);
-pub const MAX_INSTRUCTION_TRACE_LENGTH_EXCEEDED: u64 = to_builtin!(21);
-pub const BUILTIN_PROGRAMS_MUST_CONSUME_COMPUTE_UNITS: u64 = to_builtin!(22);
-pub const INVALID_ACCOUNT_OWNER: u64 = to_builtin!(23);
-pub const ARITHMETIC_OVERFLOW: u64 = to_builtin!(24);
-pub const IMMUTABLE: u64 = to_builtin!(25);
-pub const INCORRECT_AUTHORITY: u64 = to_builtin!(26);
-// Warning: Any new program errors added here must also be:
-// - Added to the below conversions
-// - Added as an equivalent to InstructionError
-// - Be featureized in the BPF loader to return `InstructionError::InvalidError`
-//   until the feature is activated
 
 impl From<ProgramError> for u64 {
     fn from(error: ProgramError) -> Self {
@@ -284,53 +254,6 @@ impl TryFrom<InstructionError> for ProgramError {
             Self::Error::Immutable => Ok(Self::Immutable),
             Self::Error::IncorrectAuthority => Ok(Self::IncorrectAuthority),
             _ => Err(error),
-        }
-    }
-}
-
-impl<T> From<T> for InstructionError
-where
-    T: ToPrimitive,
-{
-    fn from(error: T) -> Self {
-        let error = error.to_u64().unwrap_or(0xbad_c0de);
-        match error {
-            CUSTOM_ZERO => Self::Custom(0),
-            INVALID_ARGUMENT => Self::InvalidArgument,
-            INVALID_INSTRUCTION_DATA => Self::InvalidInstructionData,
-            INVALID_ACCOUNT_DATA => Self::InvalidAccountData,
-            ACCOUNT_DATA_TOO_SMALL => Self::AccountDataTooSmall,
-            INSUFFICIENT_FUNDS => Self::InsufficientFunds,
-            INCORRECT_PROGRAM_ID => Self::IncorrectProgramId,
-            MISSING_REQUIRED_SIGNATURES => Self::MissingRequiredSignature,
-            ACCOUNT_ALREADY_INITIALIZED => Self::AccountAlreadyInitialized,
-            UNINITIALIZED_ACCOUNT => Self::UninitializedAccount,
-            NOT_ENOUGH_ACCOUNT_KEYS => Self::NotEnoughAccountKeys,
-            ACCOUNT_BORROW_FAILED => Self::AccountBorrowFailed,
-            MAX_SEED_LENGTH_EXCEEDED => Self::MaxSeedLengthExceeded,
-            INVALID_SEEDS => Self::InvalidSeeds,
-            BORSH_IO_ERROR => Self::BorshIoError("Unknown".to_string()),
-            ACCOUNT_NOT_RENT_EXEMPT => Self::AccountNotRentExempt,
-            UNSUPPORTED_SYSVAR => Self::UnsupportedSysvar,
-            ILLEGAL_OWNER => Self::IllegalOwner,
-            MAX_ACCOUNTS_DATA_ALLOCATIONS_EXCEEDED => Self::MaxAccountsDataAllocationsExceeded,
-            INVALID_ACCOUNT_DATA_REALLOC => Self::InvalidRealloc,
-            MAX_INSTRUCTION_TRACE_LENGTH_EXCEEDED => Self::MaxInstructionTraceLengthExceeded,
-            BUILTIN_PROGRAMS_MUST_CONSUME_COMPUTE_UNITS => {
-                Self::BuiltinProgramsMustConsumeComputeUnits
-            }
-            INVALID_ACCOUNT_OWNER => Self::InvalidAccountOwner,
-            ARITHMETIC_OVERFLOW => Self::ArithmeticOverflow,
-            IMMUTABLE => Self::Immutable,
-            INCORRECT_AUTHORITY => Self::IncorrectAuthority,
-            _ => {
-                // A valid custom error has no bits set in the upper 32
-                if error >> BUILTIN_BIT_SHIFT == 0 {
-                    Self::Custom(error as u32)
-                } else {
-                    Self::InvalidError
-                }
-            }
         }
     }
 }
